Extract video offset detection
